backend/pyproject.toml
[project]
name = "kalshiflow"
version = "0.1.0"
description = "Real-time Kalshi trade flow visualization with WebSocket streaming"
readme = "README.md"
authors = [
    { name = "Samuel Clark", email = "sctennis77@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "cryptography>=46.0.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "starlette>=0.50.0",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
    "sortedcontainers>=2.4.0",
    # Evidence ingestion (Truth Social)
    "truthbrush>=0.2.3",
    # RL Dependencies
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "gymnasium>=0.29.0",
    "stable-baselines3>=2.0.0",
    "torch>=2.0.0",
    "scipy>=1.16.3",
    # Deep Agent Dependencies
    "anthropic>=0.30.0",
    "openai>=1.0.0",
    "duckduckgo-search>=6.0.0",
    "pyyaml>=6.0",
    # LangChain + LangGraph (arb agent orchestration)
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.2.0",
    # Supabase client for persistence
    "supabase>=2.0.0",
    # Video download for content extraction
    "yt-dlp>=2024.1.0",
    # GDELT BigQuery news intelligence
    "google-cloud-bigquery>=3.0.0",
    # Pair index: fast fuzzy string matching
    "rapidfuzz>=3.0.0",
    "httpx>=0.28.1",
]

[project.scripts]
kalshiflow = "kalshiflow:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "aiohttp>=3.13.2",
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
